requirements.txt
# pypdf2==3.0.1
# langchain==0.1.0
# sentence-transformers==2.2.2
# faiss-cpu
# torch==2.5.1
# transformers==4.36.2
# flask==3.0.0
# python-dotenv==1.0.0
# nltk==3.8.1
# tqdm==4.66.1
# psutil
# huggingface-hub==0.12.0
# accelerate


torch
transformers
sentence-transformers
faiss-cpu
langchain
flask
python-dotenv
nltk
tqdm
huggingface-hub
psutil
sentencepiece
accelerate